requirements.txt
fastapi
uvicorn
youtube-dl
yt-dlp
ffmpeg-python
smplx
transformers
speechrecognition
pydantic
pytest
python-dotenv
